# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Openlocal API
  version: 0.1.0
  description: API for Openlocal local commerce inventory, storefronts, marketplace discovery, and analytics.
servers:
  - url: http://localhost:8080
...
security:
  - clerkBearer: []
components:
  securitySchemes:
    clerkBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    BusinessIDQuery:
      name: business_id
      in: query
      required: true
      schema:
        type: string
        format: uuid
    LimitQuery:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        format: int32
        default: 25
        minimum: 1
        maximum: 100
    OffsetQuery:
      name: offset
      in: query
      required: false
      schema:
        type: integer
        format: int32
        default: 0
        minimum: 0
  schemas:
...
    Business:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        clerk_org_id:
          type: string
          description: Only returned on authenticated private business responses.
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        business_type:
          type: string
        phone:
          type: string
        whatsapp:
          type: string
        email:
          type: string
          format: email
        website:
          type: string
          format: uri
        logo_url:
          type: string
          format: uri
        cover_image_url:
          type: string
          format: uri
        status:
          type: string
        address:
          type: string
        neighborhood:
          type: string
        city:
          type: string
        state:
          type: string
        country:
          type: string
        postal_code:
          type: string
        latitude:
          type: string
        longitude:
          type: string
        pickup_available:
          type: boolean
        delivery_available:
          type: boolean
        timezone:
          type: string
          example: America/Mexico_City
          description: IANA timezone used to interpret the business opening hours.
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - slug
        - description
        - business_type
        - city
        - state
        - country
        - pickup_available
        - delivery_available
        - timezone
...
    ProductRequest:
      type: object
      additionalProperties: false
      properties:
        business_id:
          type: string
          format: uuid
        category_id:
          type: string
          format: uuid
          nullable: true
        name:
          type: string
          example: Country Sourdough
        slug:
          type: string
          pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
        description:
          type: string
        brand:
          type: string
          nullable: true
        unit:
          type: string
          default: piece
        product_type:
          type: string
          enum: [stocked_product, made_to_order_product, unique_item]
          default: stocked_product
        is_handmade:
          type: boolean
        is_public:
          type: boolean
        status:
          type: string
          enum: [draft, active, archived]
          default: draft
      required:
        - business_id
        - name
        - description
    Product:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        business_id:
          type: string
          format: uuid
        category_id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        brand:
          type: string
        unit:
          type: string
        product_type:
          type: string
        is_handmade:
          type: boolean
        is_public:
          type: boolean
        status:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - slug
        - description
        - unit
        - product_type
    VariantRequest:
      type: object
      additionalProperties: false
      properties:
        business_id:
          type: string
          format: uuid
        product_id:
          type: string
          format: uuid
        sku:
          type: string
        barcode:
          type: string
          nullable: true
        internal_code:
          type: string
        name:
          type: string
        attributes:
          type: object
          additionalProperties: true
        price:
          type: string
          example: "120.00"
        cost:
          type: string
          nullable: true
          description: Private cost, only available to authorized business users.
        currency:
          type: string
          pattern: "^[A-Z]{3}$"
          default: MXN
        track_inventory:
          type: boolean
          default: true
        public_stock_status:
          type: string
          enum: [available, low_stock, out_of_stock, made_to_order, unknown]
        reorder_point:
          type: string
          default: "0"
        lead_time_days:
          type: integer
          format: int32
          minimum: 0
        status:
          type: string
          enum: [active, archived]
          default: active
      required:
        - business_id
        - product_id
        - sku
        - name
        - attributes
        - price
    Variant:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        product_id:
          type: string
          format: uuid
        business_id:
          type: string
          format: uuid
        sku:
          type: string
        barcode:
          type: string
        internal_code:
          type: string
        name:
          type: string
        attributes:
          type: object
          additionalProperties: true
        price:
          type: string
        cost:
          type: string
        currency:
          type: string
        track_inventory:
          type: boolean
        public_stock_status:
          type: string
        reorder_point:
          type: string
        lead_time_days:
          type: integer
          format: int32
        status:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - product_id
        - sku
        - name
        - price
        - currency
        - public_stock_status
    PublicProduct:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        brand:
          type: string
        unit:
          type: string
        product_type:
          type: string
        variant_id:
          type: string
          format: uuid
        sku:
          type: string
        variant_name:
          type: string
        price:
          type: string
        currency:
          type: string
        public_stock_status:
          type: string
        image_url:
          type: string
          format: uri
      required:
        - id
        - name
        - slug
        - description
        - unit
        - product_type
        - variant_id
        - sku
        - variant_name
        - price
        - currency
        - public_stock_status
    MarketplaceProduct:
      allOf:
        - $ref: "#/components/schemas/PublicProduct"
        - type: object
          additionalProperties: false
          properties:
            business_slug:
              type: string
            business_name:
              type: string
          required:
            - business_slug
            - business_name
...
paths:
...
  /api/v1/products:
    get:
      tags:
        - Catalog
      parameters:
        - $ref: "#/components/parameters/BusinessIDQuery"
        - $ref: "#/components/parameters/LimitQuery"
        - $ref: "#/components/parameters/OffsetQuery"
      responses:
        "200":
          description: Private product list.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      tags:
        - Catalog
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductRequest"
      responses:
        "201":
          description: Product created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /api/v1/products/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags:
        - Catalog
      responses:
        "200":
          description: Private product.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    patch:
      tags:
        - Catalog
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductRequest"
      responses:
        "200":
          description: Product updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      tags:
        - Catalog
      responses:
        "204":
          description: Product archived.
  /api/v1/products/{id}/variants:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: "#/components/parameters/BusinessIDQuery"
    get:
      tags:
        - Catalog
      description: Lists all variants for a product scoped to the caller's business.
      responses:
        "200":
          description: Private variants for the product.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Variant"
  /api/v1/variants:
    post:
      tags:
        - Catalog
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VariantRequest"
      responses:
        "201":
          description: Variant created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Variant"
  /api/v1/variants/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags:
        - Catalog
      responses:
        "200":
          description: Private variant, including cost for authorized users.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Variant"
    patch:
      tags:
        - Catalog
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VariantRequest"
      responses:
        "200":
          description: Variant updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Variant"
    delete:
      tags:
        - Catalog
      responses:
        "204":
          description: Variant archived.
  /api/v1/variants/by-barcode/{barcode}:
    get:
      tags:
        - Catalog
      parameters:
        - name: barcode
          in: path
          required: true
          schema:
            type: string
        - $ref: "#/components/parameters/BusinessIDQuery"
      responses:
        "200":
          description: Business-scoped barcode lookup.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Variant"
  /api/v1/variants/by-sku/{sku}:
    get:
      tags:
        - Catalog
      parameters:
        - name: sku
          in: path
          required: true
          schema:
            type: string
        - $ref: "#/components/parameters/BusinessIDQuery"
      responses:
        "200":
          description: Business-scoped SKU lookup.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Variant"
...
  /api/v1/public/businesses/{slug}:
    get:
      tags:
        - Public Storefront
      security: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
            pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
      responses:
        "200":
          description: Public business profile.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
  /api/v1/public/businesses/{slug}/products:
    get:
      tags:
        - Public Storefront
      security: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
            pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
        - $ref: "#/components/parameters/LimitQuery"
        - $ref: "#/components/parameters/OffsetQuery"
      responses:
        "200":
          description: Public products without private cost or stock data.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PublicProduct"
  /api/v1/marketplace/search:
    get:
      tags:
        - Marketplace
      security: []
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
        - $ref: "#/components/parameters/LimitQuery"
        - $ref: "#/components/parameters/OffsetQuery"
      responses:
        "200":
          description: Public marketplace product search.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MarketplaceProduct"
  /api/v1/marketplace/businesses:
    get:
      tags:
        - Marketplace
      security: []
      parameters:
        - name: city
          in: query
          required: false
          schema:
            type: string
        - name: min_lat
          in: query
          required: false
          schema:
            type: string
        - name: max_lat
          in: query
          required: false
          schema:
            type: string
        - name: min_lng
          in: query
          required: false
          schema:
            type: string
        - name: max_lng
          in: query
          required: false
          schema:
            type: string
        - $ref: "#/components/parameters/LimitQuery"
        - $ref: "#/components/parameters/OffsetQuery"
      responses:
        "200":
          description: Public marketplace business search.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Business"
  /api/v1/marketplace/products:
    get:
      tags:
        - Marketplace
      security: []
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
        - $ref: "#/components/parameters/LimitQuery"
        - $ref: "#/components/parameters/OffsetQuery"
      responses:
        "200":
          description: Public marketplace product search.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MarketplaceProduct"